backend/pyproject.toml
[tool.poetry]
name = "to-ascii-demo"
version = "0.0.0"
description = ""
authors = ["Milo Weinberg <iapetus011@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
to-ascii = {extras = ["speedups"], version = "^5.0.1"}
fastapi = "^0.79.0"
pydantic = "^1.9.1"
python-dotenv = "^0.20.0"
uvicorn = "^0.18.2"
websockets = "^10.3"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
